% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{iplot_bar}
\alias{iplot_bar}
\title{Bar chart}
\usage{
iplot_bar(
  coin,
  dset = "Raw",
  iCode = NULL,
  usel = NULL,
  stack_children = FALSE,
  ulabs = "uCode",
  ilabs = "iCode",
  orientation = "horizontal"
)
}
\arguments{
\item{dset}{The data set to plot.}

\item{iCode}{The selected indicator code or aggregate (does not support multiple indicators)}

\item{usel}{A character vector of unit codes to highlight on the bar chart (optional)}

\item{stack_children}{If `TRUE`, produces a stacked bar chart with any children of `iCode.`
This only works for aggregates.}

\item{ulabs}{Tick labels to display on x axis: either `"uCode"` or `"uName"`.}

\item{ilabs}{if `"iName"`, displays axis and legend using indicator names, else if `"iCode"`, uses codes.}

\item{orientation}{Controls the orientation of the plot: either `"horizontal"` or `"vertical"`.}

\item{COIN}{The COIN object, or a data frame of indicator data.}
}
\value{
Interactive bar chart generated by plotly.
}
\description{
Generates an interactive bar chart.
}
\examples{
#

}
