% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{iplot_dist}
\alias{iplot_dist}
\title{Violin plots or histograms}
\usage{
iplot_dist(coin, dset = "Raw", iCode = NULL, Level = NULL, ptype = "Violin")
}
\arguments{
\item{coin}{The coin}

\item{dset}{Name of data set}

\item{iCode}{Indicator code}

\item{Level}{Index level to target - see [COINr::get_data()]}

\item{ptype}{Either `"Violin"` or `"Histogram"`.}
}
\value{
Plotly plot
}
\description{
Violin plots or histograms
}
