% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_framework.R
\name{iplot_framework}
\alias{iplot_framework}
\title{Interactive sunburst plot of index structure}
\usage{
iplot_framework(coin, plotly_colorway = NULL)
}
\arguments{
\item{coin}{The coin object}

\item{plotly_colorway}{Optional character vector of colour codes. This is passed to
plotly's "colorway" parameter, and specifies the main colours to use in the chart.}
}
\description{
Plots the structure of the index using a sunburst plot with **plotly**.
}
