% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{iplot_radar}
\alias{iplot_radar}
\title{Radar chart}
\usage{
iplot_radar(
  coin,
  dset = "Raw",
  usel = NULL,
  Level = NULL,
  iCodes = NULL,
  addstat = "none"
)
}
\arguments{
\item{coin}{The coin object}

\item{dset}{The data set to use in the table}

\item{usel}{Character vector of unit code(s) to plot data from}

\item{Level}{The selected aggregation level to take indicator data from,
where 1 is the base indicator level, and 2, 3 etc are higher aggregation levels}

\item{iCodes}{The indicator or aggregation code(s) to plot}

\item{addstat}{Adds the statistic of the scores in each dimension as a separate trace. If `"mean"` adds the overall
mean for each dimension/indicator. If `"median"` adds the overall median.}
}
\value{
Interactive radar chart generated using plotly.
}
\description{
Generates an interactive radar chart for a specified unit or set of units.
}
\details{
This function uses **plotly** to generate a radar chart for showing one or more units, compared using a specified set of indicators.
Optionally, you can add mean/median as an extra trace. The point being to show how a particular unit compares to
its peers.
}
\examples{
#

}
