% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{iplot_scatter}
\alias{iplot_scatter}
\title{Scatter plot}
\usage{
iplot_scatter(
  coin,
  dsets,
  iCodes,
  Levels,
  axes_label = "iName",
  log_axes = c(FALSE, FALSE),
  trendline = FALSE,
  show_details = FALSE,
  marker_colour = NULL
)
}
\arguments{
\item{coin}{A coin}

\item{dsets}{Character vector with names of one or two data sets}

\item{iCodes}{Character vector with codes of one or two indicators}

\item{Levels}{Numeric vector with one or two levels}

\item{axes_label}{Either `"iName"` or `"iCode"`, optionally appended with `"+dset"` to additionally
write the data set name on the axis labels.}

\item{log_axes}{2-length logical vector specifying log (TRUE) for x and y axes respectively}

\item{trendline}{Logical: if `TRUE` adds a trendline which is calculated based on the specifications of
`log_axes`. E.g. if the x axis is log-transformed, will regress y on log(x).}

\item{show_details}{Logical: if `TRUE` shows correlation and p-value as overlaid text on the plot.}

\item{marker_colour}{Optional to specify the colour of the points.}
}
\value{
Plotly object
}
\description{
Plots two indicators against each other. Each indicator is taken from the
entries in `dsets`, `iCodes` and `Levels`. For any of these with only one entry
it is repeated for both.
}
