% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_strings}
\alias{truncate_strings}
\title{Truncate a character vector}
\usage{
truncate_strings(x, maxtot = 15, maxword = 4, at_end = "")
}
\arguments{
\item{x}{Character vector}

\item{maxtot}{Max total string length}

\item{maxword}{Max word length before truncation}

\item{at_end}{Suffix to indicate string has been truncated}
}
\value{
Truncated character vector
}
\description{
For any entries in `x` that are longer than `maxtot`, individual words in
the string will be truncated to `maxword`, and the whole string will be
truncated to have total length `maxtot` or less. Also removes small words like
"in", "and", etc.
}
