% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{iplot_bar}
\alias{iplot_bar}
\title{Bar chart}
\usage{
iplot_bar(
  coin,
  dset = "Raw",
  iCode = NULL,
  usel = NULL,
  stack_children = FALSE,
  ulabs = "uCode",
  ilabs = "iCode",
  orientation = "horizontal",
  plot_subset = NULL,
  trunc_ulabs = FALSE
)
}
\arguments{
\item{coin}{The coin object}

\item{dset}{The data set to plot.}

\item{iCode}{The selected indicator code or aggregate (does not support multiple indicators)}

\item{usel}{A character vector of unit codes to highlight on the bar chart (optional)}

\item{stack_children}{If `TRUE`, produces a stacked bar chart with any children of `iCode.`
This only works for aggregates.}

\item{ulabs}{Tick labels to display on x axis: either `"uCode"` or `"uName"`.}

\item{ilabs}{if `"iName"`, displays axis and legend using indicator names, else if `"iCode"`, uses codes.}

\item{orientation}{Controls the orientation of the plot: either `"horizontal"` or `"vertical"`.}

\item{plot_subset}{Optional number to plot top or bottom-scoring subset of units. Set to a positive integer (e.g. 10)
to plot the top 10, or a negative integer to plot the bottom (e.g. -20 plots the bottom 20).}

\item{trunc_ulabs}{Logical: if `TRUE` truncates unit labels using [truncate_strings()].}
}
\value{
Interactive bar chart generated by plotly.
}
\description{
Generates an interactive bar chart.
}
\examples{
#

}
