% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_framework.R
\name{iplot_framework}
\alias{iplot_framework}
\title{Interactive sunburst plot of index structure}
\usage{
iplot_framework(coin, seg_cols = NULL)
}
\arguments{
\item{seg_cols}{A character vector of colour codes, one for each segment in the plot. The length of this
vector must be equal to the number of segments, i.e. the sum of the number of indicators and aggregates
in each level.}

\item{COIN}{COIN object, or list with first entry is the indicator metadata, second entry is the aggregation metadata}
}
\description{
Plots the structure of the index using a sunburst plot with **plotly**.
}
