% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{iplot_scatter}
\alias{iplot_scatter}
\title{Scatter plot}
\usage{
iplot_scatter(coin, dsets, iCodes, Levels, axes_label = "iName")
}
\arguments{
\item{coin}{A coin}

\item{dsets}{Character vector with names of one or two data sets}

\item{iCodes}{Character vector with codes of one or two indicators}

\item{Levels}{Numeric vector with one or two levels}

\item{axes_label}{Either `"iName"` or `"iCode"`}
}
\value{
Plotly object
}
\description{
Plots two indicators against each other. Each indicator is taken from the
entries in `dsets`, `iCodes` and `Levels`. For any of these with only one entry
it is repeated for both.
}
